fix: make `start_child` pass driver to `gen_rpc_acceptor:start_link/2`
If, for some reason, we call `gen_rpc_acceptor_sup:start_child` with a
`{Driver, Peer}` pair that is already started, it'll try to terminate
it and start anew.  But, on this second attempt, it only passes down
the `Peer`, and there is no `start_link/1` in `gen_rpc_acceptor`.